login

+++ Domain/WorkTime/Events/Event.cs
@@ -7,7 +7,7 @@ namespace Domain.WorkTimeAggregate.Events
     public class Event
     {
         public long? Id { get; internal set; }
-        public int AggregateId { get; internal set; }
+        public long AggregateId { get; internal set; }
         public long AggregateVersion { get; internal set; }
         public DateTime Date { get; internal set; }
         public EventName EventName { get; internal set; }
@@ -16,7 +16,7 @@ namespace Domain.WorkTimeAggregate.Events
         {
         }
 
-        public Event(int aggregateId, DateTime date, EventName eventName)
+        public Event(long aggregateId, DateTime date, EventName eventName)
         {
             AggregateId = aggregateId;
             Date = date;
@@ -26,7 +26,7 @@ namespace Domain.WorkTimeAggregate.Events
 
     public static class EventDeserializationHelper
     {
-        public static void Deserialize(long? id, int aggregateId, long aggregateVersion, DateTime date,
+        public static void Deserialize(long? id, long aggregateId, long aggregateVersion, DateTime date,
             EventName eventName, Event ev)
         {
             ev.Id = id;
