+++ Domain/WorkTime/Events/MouseAction.cs
@@ -4,7 +4,7 @@ namespace Domain.WorkTimeAggregate.Events
 {
     public class MouseAction : Event
     {
-        public MouseAction(int aggregateId, DateTime date, MouseKeyboardEvent mkEvent) : base(aggregateId, date, EventName.MouseAction)
+        public MouseAction(long aggregateId, DateTime date, MouseKeyboardEvent mkEvent) : base(aggregateId, date, EventName.MouseAction)
         {
             MkEvent = mkEvent;
         }
